move transaction filter check in transaction utils
